requirements.txt
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pypdf==4.3.1
PyYAML==6.0.2
